requirements.txt
django==3.2
djangorestframework
drf-yasg
django-filter

